requirements.txt
absl-py==2.1.0
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cron-descriptor==1.4.5
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.5
django-celery-beat==2.7.0
django-environ==0.12.0
django-timezone-field==7.1
flatbuffers==24.12.23
fonttools==4.55.5
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
gunicorn==23.0.0
h5py==3.12.1
html5lib==1.1
idna==3.10
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
kombu==5.4.2
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
multitasking==0.0.11
namex==0.0.8
numpy==1.26.0  # Keeping numpy version as per requirement
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==5.29.3
python-dotenv
psycopg2-binary==2.9.10
Pygments==2.19.1
pyparsing==3.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0  # Use this instead of tensorflow_intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.8.2
wrapt==1.17.2
yfinance==0.2.52
